Increase size of column for datasets zipfile_size

+++ database/migrations/2020_11_05_152811_add_zipfile_globus_columns_to_datasets.php
@@ -14,7 +14,7 @@ class AddZipfileGlobusColumnsToDatasets extends Migration
     public function up()
     {
         Schema::table('datasets', function (Blueprint $table) {
-            $table->unsignedInteger('zipfile_size')->nullable();
+            $table->unsignedBigInteger('zipfile_size')->nullable();
             $table->boolean('globus_path_exists')->nullable();
         });
     }
